started conversion script for robosuite v1.3.1. WIP
