+++ robomimic/envs/env_robosuite.py
@@ -6,7 +6,6 @@ with metadata present in datasets.
 import json
 import numpy as np
 from copy import deepcopy
-import xml.etree.ElementTree as ET
 
 import mujoco_py
 import robosuite
@@ -50,7 +49,7 @@ class EnvRobosuite(EB.EnvBase):
         # robosuite version check
         self._is_v1 = (robosuite.__version__.split(".")[0] == "1")
         if self._is_v1:
-            assert (robosuite.__version__.split(".")[1] in ["2", "3"]), "only support robosuite v0.3 and v1.2+"
+            assert (int(robosuite.__version__.split(".")[1]) >= 2), "only support robosuite v0.3 and v1.2+"
 
         kwargs = deepcopy(kwargs)
 
@@ -133,11 +132,6 @@ class EnvRobosuite(EB.EnvBase):
         if "model" in state:
             self.reset()
             xml = postprocess_model_xml(state["model"])
-
-            # ensure xml string is compatible with robosuite v1.3
-            if robosuite.__version__.split(".")[1] in ["3"]:
-                xml = postprocess_xml_for_v_1_3(xml)
-
             self.env.reset_from_xml_string(xml)
             self.env.sim.reset()
             if not self._is_v1:
@@ -201,7 +195,8 @@ class EnvRobosuite(EB.EnvBase):
                 # ensures that we don't accidentally add robot wrist images a second time
                 pf = robot.robot_model.naming_prefix
                 for k in di:
-                    if k.startswith(pf) and (k not in ret) and (not k.endswith("proprio-state")):
+                    if k.startswith(pf) and (k not in ret) and \
+                            (not k.endswith("proprio-state")) and (k in ObsUtils.OBS_KEYS_TO_MODALITIES):
                         ret[k] = np.array(di[k])
         else:
             # minimal proprioception for older versions of robosuite
@@ -370,70 +365,3 @@ class EnvRobosuite(EB.EnvBase):
         Pretty-print env description.
         """
         return self.name + "\n" + json.dumps(self._init_kwargs, sort_keys=True, indent=4)
-
-
-def postprocess_xml_for_v_1_3(xml_str):
-    """
-    Postprocess xml string generated by robosuite to be compatible with robosuite v1.3
-    This script should not the xml string if it was already generated using robosuite v1.3
-    Args:
-        xml_str (str): xml string to process (from robosuite v1.2)
-    """
-
-    def get_eef_panda_element(prefix):
-        """
-        Helper function for new element added in robosuite v1.3
-        """
-
-        eef_panda = """
-        <body name="eef" pos="0 0 0.097" quat="1 0 0 0">
-            <site name="grip_site" pos="0 0 0" size="0.01 0.01 0.01" rgba="1 0 0 0.5" type="sphere" group="1"/>
-            <site name="ee_x" pos="0.1 0 0" size="0.005 .1"  quat="0.707105  0 0.707108 0 " rgba="1 0 0 0" type="cylinder" group="1"/>
-            <site name="ee_y" pos="0 0.1 0" size="0.005 .1" quat="0.707105 0.707108 0 0" rgba="0 1 0 0" type="cylinder" group="1"/>
-            <site name="ee_z" pos="0 0 0.1" size="0.005 .1" quat="1 0 0 0" rgba="0 0 1 0" type="cylinder" group="1"/>
-            <!-- This site was added for visualization. -->
-            <site name="grip_site_cylinder" pos="0 0 0" size="0.005 10" rgba="0 1 0 0.3" type="cylinder" group="1"/>
-        </body>
-        """
-
-        tree = ET.ElementTree(ET.fromstring(eef_panda))
-        root = tree.getroot()
-        for elem in root.iter():
-            elem.attrib["name"] = "{}_{}".format(prefix, elem.get("name"))
-
-        return tree
-
-    root = ET.fromstring(xml_str)
-    worldbody = root.find('worldbody')
-    bodies_to_process = []
-    # remove elements that are specific to v1.2
-    for body in worldbody.iter('body'):
-        for site in body.findall('site'):
-            site_name = site.get('name')
-
-            # remove grip_site and grip_site_cylinder
-            if 'grip_site' in site_name:
-                # print("removing {site} from {body}".format(
-                #     site=site_name,
-                #     body=body.get('name'),
-                # ))
-                body.remove(site)
-                if body not in bodies_to_process:
-                    bodies_to_process.append(body)
-
-            # remove site called robot{}_ee{}
-            if site_name.startswith('robot') and 'ee' in site_name:
-                # print("removing {site} from {body}".format(
-                #     site=site_name,
-                #     body=body.get('name'),
-                # ))
-                body.remove(site)
-
-    # add elements introduced in v1.3
-    for body in bodies_to_process:
-        prefix = body.get("name").split('_')[0]
-        eef_tree = get_eef_panda_element(prefix)
-        body.append(eef_tree.getroot())
-
-    xml_str = ET.tostring(root, encoding="utf8").decode("utf8")
-    return xml_str
